Equinor/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
altair==4.2.0
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
azure-core==1.24.2
azure-storage-blob==12.13.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bertopic==0.11.0
bleach==5.0.1
blinker==1.4
blis==0.7.8
cachetools==5.2.0
catalogue==2.0.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
commonmark==0.9.1
cryptography==37.0.4
cssselect==1.1.0
cycler==0.11.0
cymem==2.0.6
Cython==0.29.30
datasets==2.3.2
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.5.1
emoji==1.7.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.0/en_core_web_sm-3.4.0-py3-none-any.whl
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.7.1
fonttools==4.33.3
frozenlist==1.3.0
fsspec==2022.5.0
GetOldTweets3==0.0.11
gitdb==4.0.9
GitPython==3.1.27
hdbscan==0.8.28
huggingface-hub==0.8.1
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.8.0
ipykernel==6.15.0
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
isodate==0.6.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.7.2
jupyter-client==7.3.4
jupyter-core==4.11.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
kiwisolver==1.4.3
langcodes==3.3.0
llvmlite==0.38.1
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
# mistune==2.0.3
msrest==0.7.1
multidict==6.0.2
multiprocess==0.70.13
murmurhash==1.0.7
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
nltk==3.7
notebook==6.4.12
numba==0.55.2
numpy==1.22.4
oauthlib==3.2.0
onnx==1.12.0
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
pathy==0.6.2
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
plotly==5.9.0
preshed==3.0.6
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py3langid==0.2.2
pyarrow==8.0.0
pycparser==2.21
pydantic==1.9.1
pydeck==0.7.1
Pygments==2.12.0
Pympler==1.0.1
pynndescent==0.5.7
pyparsing==3.0.9
pyquery==1.4.3
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.4.1
pyzmq==23.2.0
regex==2022.6.2
requests==2.28.1
requests-oauthlib==1.3.1
responses==0.18.0
rich==12.5.1
scikit-learn==1.1.1
scipy==1.8.1
semver==2.13.0
Send2Trash==1.8.0
sentence-transformers==2.2.2
sentencepiece==0.1.96
six==1.16.0
sklearn==0.0
smart-open==5.2.1
smmap==5.0.0
snscrape==0.4.3.20220106
soupsieve==2.3.2.post1
spacy==3.4.0
spacy-legacy==3.0.9
spacy-loggers==1.0.2
srsly==2.4.3
stack-data==0.3.0
statsmodels==0.13.2
streamlit==1.11.0
supersmoother==0.4
tdqm==0.0.1
tenacity==8.0.1
terminado==0.15.0
thinc==8.1.0
threadpoolctl==3.1.0
tinycss2==1.1.1
tokenizers==0.12.1
toml==0.10.2
toolz==0.12.0
# torch==1.12.0
# torchaudio==0.12.0
# torchvision==0.13.0
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
transformers==4.20.1
tweepy==4.10.0
typer==0.4.2
typing_extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
umap-learn==0.5.3
urllib3==1.26.9
validators==0.20.0
wasabi==0.9.1
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.6.1
xxhash==3.0.0
yarl==1.7.2
zipp==3.8.1
